requirements.txt
django==2.2.16
djangorestframework==3.12.4
django_base64field==1.0
django-filter==21.1
django-colorfield==0.7.2
six==1.16.0
pybase64==1.2.3
uuid==1.30
drf_extra_fields==3.4.0
fpdf==1.7.2
python-dotenv==0.21.0
gunicorn==20.0.4
psycopg2-binary==2.8.6